% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{distributeTopics}
\alias{distributeTopics}
\title{Distribute topics from LDA models onto new corpora}
\usage{
distributeTopics(articles, mongo)
}
\arguments{
\item{articles}{Used to create corpora.}

\item{mongo}{A \link[rmongodb]{mongo} connection.}
}
\value{
A list
}
\description{
Using an iterator from new corpora, a document-term matrix compatible with existing LDA models is
created. The \code{transform} method from the \link[text2vec]{LatentDirichletAllocation} handles the rest

The iterator from new corpora must be transformed to take into account collocations present in the
LDA models. Moreover, any terms present in the new corpora's document-term matrix, not present in the LDA models
must be dropped.
}
