% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topics.R
\name{getTopicLabel}
\alias{getTopicLabel}
\title{Label LDA topics}
\usage{
getTopicLabel(lambda = 0.3, lda, word_vector)
}
\arguments{
\item{lambda}{A numeric from 0 to 1. This determines the relevancy score}

\item{lda}{A pre-trained \link[text2vec]{LatentDirichletAllocation} model.}

\item{word_vector}{A numeric matrix. Column-bound word vectors.}
}
\value{
A character vector
}
\description{
Use vector algebra on word embeddings of most relevant words in topics to derive single topic label
}
